third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-stretch-expected.txt

This is a testharness.js-based test.
PASS Can set 'font-stretch' to CSS-wide keywords: initial
PASS Can set 'font-stretch' to CSS-wide keywords: inherit
PASS Can set 'font-stretch' to CSS-wide keywords: unset
PASS Can set 'font-stretch' to CSS-wide keywords: revert
PASS Can set 'font-stretch' to var() references:  var(--A)
FAIL Can set 'font-stretch' to the 'normal' keyword: normal assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'ultra-condensed' keyword: ultra-condensed assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'extra-condensed' keyword: extra-condensed assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'condensed' keyword: condensed assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'semi-condensed' keyword: semi-condensed assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'semi-expanded' keyword: semi-expanded assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'expanded' keyword: expanded assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'extra-expanded' keyword: extra-expanded assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'ultra-expanded' keyword: ultra-expanded assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
PASS Can set 'font-stretch' to a percent: 0%
FAIL Can set 'font-stretch' to a percent: -3.14% assert_approx_equals: expected -3.14 +/- 0.000001 but got -3
FAIL Can set 'font-stretch' to a percent: 3.14% assert_approx_equals: expected 3.14 +/- 0.000001 but got 3
PASS Can set 'font-stretch' to a percent: calc(0% + 0%)
PASS Setting 'font-stretch' to a length throws TypeError
PASS Setting 'font-stretch' to a time throws TypeError
PASS Setting 'font-stretch' to an angle throws TypeError
PASS Setting 'font-stretch' to a flexible length throws TypeError
PASS Setting 'font-stretch' to a number throws TypeError
PASS Setting 'font-stretch' to a URL throws TypeError
PASS Setting 'font-stretch' to a transform throws TypeError
Harness: the test ran to completion.
